Fixed Home State Issue After Logout
