+++ app/backend/Controllers/HomeController.cs
@@ -137,7 +137,7 @@ public class HomeController : Controller
         }
         catch(Exception e) {
             await transaction.RollbackAsync();
-            return StatusCode(500, new { error = "Failed to create team", details = e.Message });
+            return StatusCode(500, new { error = "Failed to update activity", details = e.Message });
         }
         return StatusCode(201, new { message = "User activity was updated." });
     }
