+++ app/frontend/src/pages/HomePage.tsx
@@ -3,16 +3,20 @@ import { Box, useMediaQuery, useTheme } from "@mui/material";
 import { useEffect, useState } from "react";
 import SideBar from "../components/Sidebar/SideBar";
 import { UserActivity, ITeamModel, IUserModel } from "../models/models";
+import Cookies from "js-cookie";
 
 import wretch from "wretch";
 import ChatArea from "../components/Chat/ChatArea";
 import { useApplicationStore } from "../stores/ApplicationStore";
+import { useNavigate } from "react-router-dom";
 import { useUserStore } from "../stores/UserStore";
 import { API_URL } from "../utils/FetchUtils";
 
 export default function HomePage() {
   const theme = useTheme();
   const isBrowser = useMediaQuery(theme.breakpoints.up("sm"));
+  const navigate = useNavigate();
+  const setIsLoggedIn = useUserStore((state) => state.setIsLoggedIn);
 
   // stores for state management
   const applicationState = useApplicationStore();
@@ -20,28 +24,33 @@ export default function HomePage() {
   const [activity, setActivity] = useState<string>(UserActivity.Online);
   const [lastUpdate, setLastUpdate] = useState<Date>();
 
-  const updateActivity = () => {
-    if(!userState.isLoggedIn || Date.now() - (lastUpdate?.getTime() ?? Date.now() - 10000) < 1000) return;
-    setActivity(UserActivity.Online);
-    activitySubmit(UserActivity.Online);
-    setLastUpdate(new Date(Date.now()));
-  }
-
   const setupActivityListeners = () => {
-    document.addEventListener("keydown", () => {updateActivity();});
-    document.addEventListener("click", () => {updateActivity();});
+    document.addEventListener("keydown", () => {activitySubmit(UserActivity.Online);});
+    document.addEventListener("click", () => {activitySubmit(UserActivity.Online);});
   };
 
   const removeActivityListeners = () => {
-    document.removeEventListener("keydown", () => {updateActivity();});
-    document.removeEventListener("click", () => {updateActivity();});
+    document.removeEventListener("keydown", () => {activitySubmit(UserActivity.Online);});
+    document.removeEventListener("click", () => {activitySubmit(UserActivity.Online);});
   };
 
   const activitySubmit = (status: string) => {
+    if(Date.now() - (lastUpdate?.getTime() ?? Date.now() - 10000) < 1000) return;
+    console.log(status);
+    setActivity(status);
+    setLastUpdate(new Date(Date.now()));
     wretch(`${API_URL}/api/home/activity`)
       .auth(`Bearer ${localStorage.getItem("jwt-token")}`)
       .post({ Activity: status })
-      .res(() => {})
+      .res(() => {
+        if(status == "Offline") {
+          window.location.reload();
+          Cookies.remove("isLoggedIn");
+          localStorage.removeItem("jwt-token");
+          setIsLoggedIn(false);
+          navigate("/login");
+        }
+      })
       .catch((error) => {
         console.error("Error submitting activity:", error);
       });
